// src/enums/finbox.enum.ts
export enum FinboxUrls {
  CREATE_URL = "/back_to_application",
  LOGO_URL = "/loanapply/public/logo.png",
}

export enum FinBoxBankConnectProgressStatus {
  PROCESSING = "processing",
  FAILED = "failed",
}

export enum DateDifference {
  LESSER = 15,
  GREATER = 36,
}

export const nameCheckPercentage = 70;